I made some changes This changes is focused to bring the last updates from the CupcakeView to rest of other views. And further of this I made some improvements in the codebase to improve even more the performance and the code style.

+++ Common/Views/Cupcake View/CupcakeView.swift
@@ -85,8 +85,12 @@ struct CupcakeView: View {
                     viewModel.insertNewCupcake(with: cupcake)
                 }
             }
-            .onChange(of: userRepo.user) { _, _ in
-                viewModel.deleteAllCupcakes(with: modelContext)
+            .onChange(of: userRepo.user) { oldValue, newValue in
+                if oldValue == nil && newValue != nil {
+                    viewModel.loadCupcakes(with: userRepo.user != nil, and: modelContext)
+                } else if newValue == nil {
+                    viewModel.deleteAllCupcakes(with: modelContext)
+                }
             }
             #endif
         }
